requirements.txt
pandas>=2.0.0
kiteconnect>=4.0.0
pytz>=2023.3
numpy>=1.24.0
python-dotenv>=1.0.0
email-validator>=2.0.0
pydantic>=2.0.0
flask>=2.3.0
gunicorn>=21.0.0